Move p2p tests to replicatio tests class
