+++ src/Couchbase.Lite.Tests.Shared/Couchbase.Lite.Tests.Shared.projitems
@@ -24,7 +24,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)LogTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MigrationTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)NotificationTest.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)P2PTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PerfTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PredictiveQueryTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)QueryTest.cs" />
